Add ChunkArray
ChunkArray is a wrapper over std::unordered_map<std::pair<int, int>,
Chunk>. Its purpose is as a container for Chunks. It is a separate class
rather than a typedef for std::unordered_map<std::pair<int, int>, Chunk>
in order to enable manipulation of Chunks via Topology once that is
implemented.