Fix compilation error in ChunkArray
std::pair normally can't be hashed and so can't be used as a key in
std::unordered_map. ChunkArray uses it as such and this was causing
a compilation error. Fix this by adding a header util.h containing
a struct pair_hash with a template operator() which simply takes an
std::pair, computes the hash of first and second, and combines them
using some magic stolen from boost. It now compiles.