Add Neighbourhood and NeighbourhoodType
The architecture is as follows:
NeighbourhoodType is mostly an abstract factory for Neighbourhood, but
it also provides details on the type level about the Neighbourhood (i.e.
the total number of cells and the affecting distance). Neighbourhood
represents the neighbourhood of an actual cell and is designed to be
moved around and between chunks. One Neighbourhood instance should be
used per generation, while one NeighbourhoodType instance should be
used for the entire run (or at least until the neighbourhood type is
changed).