Add Ruleset
Ruleset is a class which owns a NeighbourhoodType* and two dynamic
bool arrays representing whether a dead cell becomes alive, or a live
cell stays alive, with each number of possible live neighbours. These
arrays have the same size as the number of total cells in the
neighbourhood type, plus one (to accommodate both the case in which
there are 0 live neighbours and that in which all cells in the
neighbourhood are live). The memory management is implemented manually
rather than using smart pointers because I wanted to. It probably has
a bunch of memory leaks. Oh well.

Also, make NeighbourhoodType (and Neighbourhood) more polymorphism-
friendly by adding virtual default destructors to both and a virtual
clone() method to NeighbourhoodType for polymorphic cloning in Ruleset.