Implement Chunk::tick for sides of chunks
Chunk::tick now supports specifying a chunk side and an affecting
distance; if specified, it will only scan the cells within the affecting
distance of the specified side.

Add a bunch of functionality to Side to support this; in particular, add
Side::right() and Side::left(), which return the Side to the right and
left of this side respectively, viewing this side from the bottom.

Add Neighbourhood::moveUp() because it turns out that I actually do
need it for sides other than bottom. Add Neighbourhood::moveToSide(Side)
which calls the appropriate move[U/D/L/R] method based on the Side it's
given.

Make Topology's destructor virtual because oops I forgot earlier.
Make a bunch of Ruleset methods const because, again, oops I forgot.