Make Chunk use second array for next generation
Oops. Previously we were doing all updates on the same array, which
would have been very bad and not correct at all. Give Chunk a second
array, newCells_, with which to process the new generation.

In order to facilitate this, split Chunk::tick into two functions:
Chunk::generate does the majority of the work Chunk::tick did in
generating the new generation, and update() copies the new generation
to the main generation and resets the new generation array.