Add Automaton
Automaton is the master class for the model of the cellular automaton;
it contains the ChunkArray, the Ruleset, and the Topology (shared
with ChunkArray, for some reason). Automaton::tick() is the function
which actually moves the entire automaton to the next generation: this
is what we've been building towards!

Also, add ChunkArray::hasNonEmpty(int x, int y) for convenience. It
just returns whether the ChunkArray has a non-empty chunk at (x, y).