Add ChunkGraphicsItem, Chunk signals
Add ChunkGraphicsItem. It is a QGraphicsItem which draws a Chunk on the
screen. It also inherits from QObject in order to have a slot to get
updated by Chunk. ChunkGraphicsItem holds a pointer (ish) to its
corresponding Chunk's cell array.

Make Chunk a QObject in order to emit a signal when its cells change
(i.e. on update() calls). Also make ChunkArray a QObject in order to
emit signals when chunks are added or removed.

Since QObjects cannot be copied, make ChunkArray's internal
std::unordered_map hold Chunk* instead of Chunk. This also necessitates
deleting the chunk pointers in ChunkArray's destructor and the
erase() method.