Add AutomatonScene and click handling
AutomatonScene is the QGraphicsScene subclass; its purpose is to
handle mouse events.

On click, make AutomatonScene calculate the chunk and cell coordinates
on which the user clicked and flip the appropriate cell. Then,
update the display.

Also, add Chunk::setCell (because apparently we didn't have that before)
and fix a misunderstanding in ChunkGraphicsItem::boundingRect().

(New favourite euphemism: it's not a bug, just a "misunderstanding".)

(Note: this is the first actual tested graphical thing in this project!)