Fix bug with removing ChunkGraphicItems twice
In MainWindow's destructor, we previously deleted the ChunkGraphicItems,
then deleted the Automaton. When the Automaton was deleted, so was the
ChunkArray, which deleted each Chunk and emitted a signal for each one
which was received by MainWindow, which deleted the item which we had
previously deleted, resulting in an error. Fix: don't delete the items
the first time (instead replace with an explanatory comment).