Allow changing automaton rules
Add RulesDialog, which the user can spawn via Settings>Change Rules
or Ctrl/Cmd-R. It displays a bunch of checkboxes which can be used to
control the rules of the automaton. (Note: it cannot yet be used to
change the topology or the neighbourhood type.)

We now support all 2D Life-like automata based on the Moore
neighbourhood!